Extend the VL<nf>R and VS<nf>R instruction, when <nf> is 2, 4, 8.
These instruction are added back in the risc-v vector spec,

commit 0fec0e257cca5e389a0e1774e7d5aaac729b4ad2,
Added clarifications on loads/stores.

Note that the base V extension only enable the single register load/store
variant with <nf> is 0.  It is clear to me that how to enable other register
load/store is <nf> is 2, 4, 8, so I will add the limitation when it is
clarified.